refactor(version): refactor version command
- Refactored version command to separate CLI handling from functions.
- Updated pyproject.toml and uv.lock to reflect version.

Version number bumped to: 0.6.3
