+++ src/evaluator/commands/version.py
@@ -1,33 +1,13 @@
 '''
-=======================================
-EValuator: PRINT VERSION
-=======================================
+Commands: version
 '''
-# ====================
-# Import external dependencies
-# ====================
+# -- Import external dependencies ----------------
 import tomllib, typer
 from rich import print
 
-# ====================
-# Initialise typer as evaluatorVersion
-# ====================
-evaluatorVersion = typer.Typer(
-    # Disable --install-completion and --show-completion options in terminal
-    add_completion=False,
-    # Disable --help option in terminal
-    add_help_option=False
-)
-
-# ====================
-# Define command: version
-# ====================
-@evaluatorVersion.command(rich_help_panel="Utility Commands")
-def version():
-    '''
-    Print current EValuator version to terminal and exit.
-    '''
+# -- Define printVersion command -----------------
+def printVersion():
     with open('pyproject.toml', 'rb') as f:
-       contents = tomllib.load(f)
+        contents = tomllib.load(f)
     print(f"\nRunning EValuator version: v{contents['project']['version']}\n")
     typer.Exit(0)
